Standard chemical potentials are updated at every iteration.
This update at every iteration is needed because there are some
standardthermodynamic models in which the standard properties
arecorrected based on composition, which changes every iteration. For
example, when using Phreeqc as a thermodynamic backend,log(K) of
reactions are corrected for ionic strength, and not only for temperature
and pressure.