PYTHON_INSTALL_PREFIX if not found, set to CMAKE_INSTALL_PREFIX
The reason for this change is to be more uniform with the default paths
for both C++ and Python outputs. CMAKE_INSTALL_PREFIX should control
install path to both output types. If a different prefix is needed for
Python package, then the user can set the variable
PYTHON_INSTALL_PREFIX.