Improvement in CMAKE and conda integration in Linux (#32)
* Added cmake modules ColoredText and ParallelJobs

* Updated pugixml from 1.6 to 1.9

* Using PROJECT_VERSION cmake variable instead of REAKTORO_VERSION

* Improved comments and include directory paths

* Forcing find_package to find pybind11 installed by cmake

* PYTHON_INSTALL_PREFIX if not found, set to CMAKE_INSTALL_PREFIX

The reason for this change is to be more uniform with the default paths
for both C++ and Python outputs. CMAKE_INSTALL_PREFIX should control
install path to both output types. If a different prefix is needed for
Python package, then the user can set the variable
PYTHON_INSTALL_PREFIX.

* THIRDPARTY_TARGETS no longer needed and removed.

The call `add_dependencies(ReaktoroObject ${THIRDPARTY_TARGETS})` is no
longer needed because the thirdparty (dependencies) targets are now
built in an explicit pre-build step, and thus this is not needed
anymore.

Note that cmake variable REAKTORO_SOURCE_DIR is now replaced by
PROJECT_SOURCE_DIR, which is already defined by cmake.

* Added # -*- coding: utf-8 -*- to avoid SyntaxError: Non-ASCII character

* Update pugixml from 1.6 to 1.9

* Delete pugixml 1.6

* Delete pugixml 1.9 in the wrong dir

* Add pybind11 to be installed on the fly with ExternalProject_Add

* Replaced word Boost by pybind11 in comment

* Define a command to install the thirdparty deps on the fly.

The installation of the external deps now occur during configure time.

Note that many cmake variables were renamed, now with prefix REAKTORO_,
for example, REAKTORO_BUILD_PYTHON instead of BUILD_PYTHON.

* Using REAKTORO_THIRDPARTY_INSTALL_ARGS instead of THIRDPARTY_COMMON_ARGS

* Implement cmake scripts for installation of Reaktoro and its deps.

* Using REAKTORO_THIRDPARTY_INSTALL_PATH instead of
REAKTORO_DEPENDENCIES_PATH

* Improve name of cmake variables.

* Rolling back some changes to make it less distant from before.

In the end, this is just some minor corrections and improvements in
cmake usage. The scripts install are not used anymore (they can be
removed or adapted to this change)

* Improving install and creation/setup of libraries to avoid duplication

* Changed CMAKE_SOURCE_DIR to PROJECT_SOURCE_DIR

* CMAKE_INSTALL_PREFIX is now used to install the python package.

* Removing the use of GNUInstallDirs for thirdparty deps installation

* Fix mismatch between default python version and the one found by cmake.

With this change, if a python version is not given using either
PYBIND11_PYTHON_VERSION or PYTHON_EXECUTABLE options, then the version
of the default python in the system (or environment) is selected (that
is, the version obtained by running `python --version`).

* Recognize if a conda environment is active.

If so, then CMAKE_INSTALL_PREFIX is set to env variable CONDA_PREFIX.

With this, a `make install` command installs all targets to the conda
environment, instead of system environment.

* Change CondaRecognizer to change CMAKE_INSTALL_PREFIX only if it was not overriden by the user

* Improved logic and BOOST_ROOT now set to CONDA_PREFIX

* Using ccache to speed up compilation

* Using REAKTORO_PYTHON_INSTALL_PREFIX

* Added ccache in the dependencies.

* Fixing CI build issues:

This should fix:
~~~
CMake Warning:
  Manually-specified variables were not used by the project:
    BOOST_INCLUDE_DIR
    BUILD_ALL
    CMAKE_VERBOSE_MAKEFILE
    PYTHON_INSTALL_PREFIX
~~~

* Checking if linux or osx to add ccache dependency.

* ccache seems to have no effect in travis and appveyor.

- the cache seems to be lost in every travis build.
- ccache is not available in win, it should not be used with appveyor
- compilation with ccache is slow for the first time, and quick
  subsequently, but since the cache is lost, it ends up slowing down
  the build

* Using cache: ccache option in travis

Suggested by @tadeu, reference:
https://docs.travis-ci.com/user/caching/#ccache-cache

* Added pugixml as dependency in conda env

* Following travis guidelines to enable ccache

* Fixed logic: and => or (linux or mac)

* Enabling ccache again to see if travis works with ccache