FormationReaction no longer has details about the product species.
This simplification avoids duplicated info from parent components. For example, a Species knows its name. A FormationReaction for it does not need to have the species name. Otherwise, in the database, we need to have something like this:

Name: Calcite
Formula: CaCO3
AggregateState: Solid
StandardThermoModel:
  FormationReaction:
    Product: Calcite
    Reactants: { Ca++: 1.0, CO3--: 1.0 }
    ReactionThermoModel:
      VantHoff:
        lgKr: 12.445
        dHr: 2323

where Calcite is repeated in the FormationReaction block