Ensure at least one Newton iteration is applied in kinetics calculations.
For kinetics, it is crucial that every calculation performs at least one Newton step, even if the 0th iteration already shows sactisfatory residuals that pass the convergence test. Without this, we see simulations in which, for example, a mineral initially absent (with amount 1e-16) that should precipitate ends up not precipitating because the calculation already exists at the 0th iteration, without a Newton step applied.