Fix bug with static variables in activity models
Some variables where declared static in the Duan-Sun and Rumpf activity
models, while they where never invalidated. Therefore, running multiple
calculations in sequence for different chemical systems and states, led
to a misuse of cached values.

An auxiliary `AqueousMixture` object is created ouside of the lambda and
the precalculated indices are captured by value to retain the original
caching intent.